chore(web): move utils from `index.ts` to separate files
