+++ packages/web/src/common/utils/round/round.util.ts
@@ -3,13 +3,3 @@ export const roundToNext = (number: number, roundBy: number): number =>
 
 export const roundToPrev = (number: number, roundBy: number): number =>
   Math.floor(number / roundBy) * roundBy;
-
-export const sortLowToHigh = (vals: number[]) =>
-  [...vals].sort(function (a, b) {
-    return a - b;
-  });
-
-export const sum = (arr: number[]) =>
-  arr.reduce(function (a, b) {
-    return a + b;
-  }, 0);
