+++ packages/web/src/views/Calendar/components/Event/Grid/GridEventPreview/snap.grid.ts
@@ -1,4 +1,4 @@
-import { roundToPrev } from "@web/common/utils";
+import { roundToPrev } from "@web/common/utils/round/round.util";
 import { Measurements_Grid } from "@web/views/Calendar/hooks/grid/useGridLayout";
 import { GRID_TIME_STEP } from "@web/views/Calendar/layout.constants";
