+++ packages/web/src/views/Calendar/hooks/grid/useDateCalcs.ts
@@ -5,7 +5,7 @@ import weekPlugin from "dayjs/plugin/weekOfYear";
 import { MutableRefObject } from "react";
 import { HOURS_AM_FORMAT } from "@core/constants/date.constants";
 import { ACCEPTED_TIMES } from "@web/common/constants/web.constants";
-import { roundToPrev } from "@web/common/utils";
+import { roundToPrev } from "@web/common/utils/round/round.util";
 import { Measurements_Grid } from "@web/views/Calendar/hooks/grid/useGridLayout";
 import { GRID_TIME_STEP } from "@web/views/Calendar/layout.constants";
 import { GRID_X_START } from "@web/views/Calendar/layout.constants";
